+++ Bit/src/Bit/Core/TimeStamp.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Bit/Core/Defines.h"
 #include <chrono>
 
 
@@ -9,8 +10,8 @@ namespace BitEngine
 class Time
 {
 private:
-    float m_LastFrameTime = 0.0f;
-    float m_DeltaTime = 0.0f;
+    f64 m_LastFrameTime = 0.0f;
+    f64 m_DeltaTime = 0.0f;
     
 public:
     void Reset()
@@ -20,7 +21,7 @@ public:
     
     void Update()
     {
-        float currentTime = GetCurrentTime();
+        f64 currentTime = GetCurrentTime();
         m_DeltaTime = currentTime - m_LastFrameTime;
         m_LastFrameTime = currentTime;
         
@@ -29,14 +30,14 @@ public:
             m_DeltaTime = 0.1f;
     }
     
-    float GetDeltaTime() const { return m_DeltaTime; }
+    f64 GetDeltaTime() const { return m_DeltaTime; }
     
 private:
-    float GetCurrentTime()
+    f64 GetCurrentTime()
     {
         static auto startTime = std::chrono::high_resolution_clock::now();
         auto currentTime = std::chrono::high_resolution_clock::now();
-        return std::chrono::duration<float>(currentTime - startTime).count();
+        return std::chrono::duration<f64>(currentTime - startTime).count();
     }
 };
 }
